Introduce distinct semantics for string readers and string converters
When converting from string kind string, we possibly want to bypass
the reader. The reason is, that the string we are given has already
been selected, and we just need to bless it as a string, while in a
reader we consider the string to be read is within a flow of other
elements, therefore the normal scanning conventions should apply.

Therefore, strings starting with a '"' are identically processed by
reader_kind, reader_same, of_string, of_string_kind and
of_string_same.

Strings not starting with a '"' are processed with the normal scanning
convention by reader_kind and reader_same while they are processed as
a whole by of_string, of_string_kind and of_string_same.  The normal
scanning convention to read a string is to stop at the first
whitespace or at the end of input, whichever condition comes first.